+++ packages/terra-clinical-data-grid/src/terra-dev-site/test/clinical-data-grid/NoPinnedColumnDataGrid.test.jsx
@@ -82,6 +82,7 @@ class NoPinnedColumnDataGrid extends React.Component {
     };
   }
 
+  /* eslint-disable react/forbid-dom-props */
   render() {
     return (
       <div id="no-pinned-column-data-grid" style={{ height: '100%', width: '100%' }}>
@@ -97,6 +98,7 @@ class NoPinnedColumnDataGrid extends React.Component {
       </div>
     );
   }
+  /* eslint-enable react/forbid-dom-props */
 }
 
 export default NoPinnedColumnDataGrid;