+++ packages/terra-clinical-data-grid/src/terra-dev-site/test/clinical-data-grid/SelectableDataGrid.test.jsx
@@ -91,6 +91,7 @@ class SelectableDataGrid extends React.Component {
     };
   }
 
+  /* eslint-disable react/forbid-dom-props */
   render() {
     const { columns } = this.state;
 
@@ -117,12 +118,12 @@ class SelectableDataGrid extends React.Component {
           onColumnSelect={(columnId) => {
             const newColumns = {};
 
-            const columnToSort = Object.assign({}, this.state.columns[columnId]);
+            const columnToSort = { ...this.state.columns[columnId] };
             columnToSort.sortIndicator = columnToSort.sortIndicator === 'ascending' ? 'descending' : 'ascending';
             newColumns[`${columnId}`] = columnToSort;
 
             if (columnId !== this.state.sortedColumnId) {
-              const previouslySortedColumn = Object.assign({}, this.state.columns[this.state.sortedColumnId]);
+              const previouslySortedColumn = { ...this.state.columns[this.state.sortedColumnId] };
               if (previouslySortedColumn) {
                 previouslySortedColumn.sortIndicator = undefined;
                 newColumns[`${this.state.sortedColumnId}`] = previouslySortedColumn;
@@ -130,7 +131,7 @@ class SelectableDataGrid extends React.Component {
             }
 
             this.setState(prevState => ({
-              columns: Object.assign({}, prevState.columns, newColumns),
+              columns: { ...prevState.columns, ...newColumns },
               sortedColumnId: columnId,
               sortedColumnDirection: columnToSort.sortIndicator,
             }));
@@ -159,6 +160,7 @@ class SelectableDataGrid extends React.Component {
       </div>
     );
   }
+  /* eslint-enable react/forbid-dom-props */
 }
 
 export default SelectableDataGrid;