+++ packages/terra-clinical-data-grid/src/terra-dev-site/test/clinical-data-grid/SubsectionDataGrid.test.jsx
@@ -102,6 +102,7 @@ class SubsectionDataGrid extends React.Component {
     };
   }
 
+  /* eslint-disable react/forbid-dom-props */
   render() {
     return (
       <div id="subsection-data-grid" style={{ height: '100%', width: '100%' }}>
@@ -126,6 +127,7 @@ class SubsectionDataGrid extends React.Component {
       </div>
     );
   }
+  /* eslint-enable react/forbid-dom-props */
 }
 
 export default SubsectionDataGrid;