+++ packages/terra-clinical-detail-view/src/terra-dev-site/test/clinical-detail-view/DetailView/DetailViewDividedSmallerTitles.test.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import DetailView from '../../../../DetailView';
+import styles from './styles.module.scss';
 
 const item = (<p>A Detail List Item</p>);
 const item2 = (<p>A Detail List Item to prove float concept when screen readjusts smaller.</p>);
@@ -11,11 +12,7 @@ const DetailViewDividedSmallerTitles = () => (
     secondaryTitles={['secondaryTitle1', 'secondaryTitle2']}
     subtitles={['subtitle1', 'subTitle2']}
     graph={(
-      <div
-        style={{
-          border: '1px solid black', height: '20em', width: '25em', marginBottom: '0.714rem',
-        }}
-      >
+      <div className={styles.placeholder}>
         This is where any visualizations would go
       </div>
 )}